# openapi.yaml
openapi: 3.1.1
info:
  title: Fake Store API
  version: 1.2.3
  description: |-
    This is a sample FakeAPI Store Server based on the OpenAPI 3.X specification. You can find out more about Swagger at https://swagger.io. In the third iteration of the pet store, we've switched to the design first approach. You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
    
    Some useful links:
    - [The Fake Store repository](https://github.com/Ismabc05/curso-api-first)
  termsOfService: http://swagger.io/terms/
  contact:
    name: Ismael Bedmar
    email: ismaelbedmarcejas@gmail.com
    url: https://github.com/Ismabc05
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
externalDocs:
  description: Find out more about Fake Store API
  url: https://github.com/Ismabc05/curso-api-first.git
tags:
  - name: Products
    description: Everything about your Products
  - name: Users
    description: Everything about your Users
servers:
  - url: http://localhost:3000/v1
    description: Version 1 de la API
  - url: http://localhost:3000/v2
    description: Version 2 de la API
components:
  schemas:
    HelloResponse:
      type: object
      properties:
        message:
          type: string
      example:
        message: "Hello world"
    Product:
      type: object
      required:
        - name
        - price
        - category
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 40
          example: "Cafetera Espresso Pro"
        description:
          type: string
          maxLength: 500
          example: "Cafetera espresso automática de 15 bares con molinillo integrado y depósito de agua de 1.5 litros."
        price:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 129.99
        category:
          type: string
          enum: [Electronics, Books, Clothes, Food]
          example: Electronics
        tags:
          type: array
          items:
            type: string
          minItems: 1
          example: ["coffee", "kitchen", "premium"]
        inStock:
          type: boolean
          example: true
        specifications:
          type: object
          additionalProperties:
            type: string
          example:
            power: "1450W"
            capacity: "1.5L"
            pressure: "15 bar"
        ratings:
          type: array
          items:
            type: object
            required:
              - score
              - comment
            properties:
              score:
                type: integer
                minimum: 1
                maximum: 5
              comment:
                type: string
                maxLength: 200
          example:
            - score: 5
              comment: "Excelente calidad de café y muy fácil de usar."
            - score: 4
              comment: "Buen producto, aunque un poco ruidoso al moler."
      example:
        name: "Cafetera Espresso Pro"
        description: "Cafetera espresso automática de 15 bares con molinillo integrado y depósito de agua de 1.5 litros."
        price: 129.99
        category: Electronics
        tags:
          - "coffee"
          - "kitchen"
          - "premium"
        inStock: true
        specifications:
          power: "1450W"
          capacity: "1.5L"
          pressure: "15 bar"
        ratings:
          - score: 5
            comment: "Excelente calidad de café y muy fácil de usar."
          - score: 4
            comment: "Buen producto, aunque un poco ruidoso al moler."
    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "joe@example.com"
        password:
          type: string
          minLength: 6
          example: "password123"
    AuthResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Autenticación mediante token JWT. Debe incluirse en el encabezado Authorization."
security:
  - JWT: []
paths:
  /hello:
    get:
      security: []
      summary: Hello World
      responses:
        '200':
          description: A succesful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloResponse'
  /auth/login:
    post:
      tags:
        - Users
      security: []
      summary: Autenticar usuario y obtener token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
            examples:
              loginCredentials:
                summary: Credenciales válidas de usuario
                value:
                  email: "joe@example.com"
                  password: "password123"
      responses:
        '200':
          description: Token JWT generado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                authSuccess:
                  summary: Autenticación exitosa
                  value:
                    token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                invalidCredentials:
                  summary: Credenciales inválidas
                  value:
                    message: "Email o contraseña incorrectos"
  /products:
    post:
      tags:
        - Products
      summary: Crear un nuevo producto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
            examples:
              newProduct:
                summary: Nuevo producto de tienda real
                value:
                  name: "Audífonos inalámbricos NoiseFree"
                  description: "Audífonos bluetooth con cancelación de ruido, batería de 30 horas y diseño ergonómico."
                  price: 79.99
                  category: Electronics
                  tags:
                    - "audio"
                    - "wireless"
                    - "travel"
                  inStock: true
                  specifications:
                    battery: "30h"
                    color: "Negro"
                    connectivity: "Bluetooth 5.2"
                  ratings:
                    - score: 5
                      comment: "Muy cómodos y el sonido es claro."
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              examples:
                createdProduct:
                  summary: Producto creado exitosamente
                  value:
                    name: "Audífonos inalámbricos NoiseFree"
                    description: "Audífonos bluetooth con cancelación de ruido, batería de 30 horas y diseño ergonómico."
                    price: 79.99
                    category: Electronics
                    tags:
                      - "audio"
                      - "wireless"
                      - "travel"
                    inStock: true
                    specifications:
                      battery: "30h"
                      color: "Negro"
                      connectivity: "Bluetooth 5.2"
                    ratings:
                      - score: 5
                        comment: "Muy cómodos y el sonido es claro."
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                invalidProduct:
                  summary: Ejemplo de error de validación
                  value:
                    message: "El campo price debe ser un número mayor o igual a 0."
    get:
      tags:
        - Products
      summary: Obtener todos los productos
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
              examples:
                productList:
                  summary: Lista de productos disponibles
                  value:
                    - name: "Audífonos inalámbricos NoiseFree"
                      description: "Audífonos bluetooth con cancelación de ruido, batería de 30 horas y diseño ergonómico."
                      price: 79.99
                      category: Electronics
                      tags:
                        - "audio"
                        - "wireless"
                        - "travel"
                      inStock: true
                      specifications:
                        battery: "30h"
                        color: "Negro"
                        connectivity: "Bluetooth 5.2"
                      ratings:
                        - score: 5
                          comment: "Muy cómodos y el sonido es claro."
                    - name: "Mochila Urbana Flex"
                      description: "Mochila resistente al agua con compartimento para portátil y múltiples bolsillos organizadores."
                      price: 59.9
                      category: Clothes
                      tags:
                        - "travel"
                        - "office"
                      inStock: true
                      specifications:
                        material: "Nylon"
                        capacity: "20L"
                        color: "Gris"
                      ratings:
                        - score: 4
                          comment: "Buena capacidad y estilo moderno."
  /products/{id}:
    get:
      tags:
        - Products
      summary: Obtener producto por id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              examples:
                foundProduct:
                  summary: Producto encontrado
                  value:
                    name: "Audífonos inalámbricos NoiseFree"
                    description: "Audífonos bluetooth con cancelación de ruido, batería de 30 horas y diseño ergonómico."
                    price: 79.99
                    category: Electronics
                    tags:
                      - "audio"
                      - "wireless"
                      - "travel"
                    inStock: true
                    specifications:
                      battery: "30h"
                      color: "Negro"
                      connectivity: "Bluetooth 5.2"
                    ratings:
                      - score: 5
                        comment: "Muy cómodos y el sonido es claro."
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                notFound:
                  summary: Producto no encontrado
                  value:
                    message: "Producto no encontrado"
    put:
      summary: Actualizar producto por id
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
            examples:
              updateProduct:
                summary: Actualizar producto
                value:
                  name: "Audífonos inalámbricos NoiseFree"
                  description: "Audífonos bluetooth con cancelación de ruido mejorados con micrófono."
                  price: 84.99
                  category: Electronics
                  tags:
                    - "audio"
                    - "wireless"
                    - "travel"
                  inStock: true
                  specifications:
                    battery: "30h"
                    color: "Negro"
                    connectivity: "Bluetooth 5.2"
                  ratings:
                    - score: 5
                      comment: "Mejorados tras la última actualización."
      responses:
        '200':
          description: Producto actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              examples:
                updatedProduct:
                  summary: Producto actualizado
                  value:
                    name: "Audífonos inalámbricos NoiseFree"
                    description: "Audífonos bluetooth con cancelación de ruido mejorados con micrófono."
                    price: 84.99
                    category: Electronics
                    tags:
                      - "audio"
                      - "wireless"
                      - "travel"
                    inStock: true
                    specifications:
                      battery: "30h"
                      color: "Negro"
                      connectivity: "Bluetooth 5.2"
                    ratings:
                      - score: 5
                        comment: "Mejorados tras la última actualización."
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                notFound:
                  summary: Producto no encontrado
                  value:
                    message: "Producto no encontrado"
    delete:
      summary: Eliminar producto por id
      tags:
        - Products
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Producto eliminado
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                notFound:
                  summary: Producto no encontrado
                  value:
                    message: "Producto no encontrado"
  /users:
    post:
      tags:
        - Users
      summary: Crear un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              require:
                - name
                - age
                - email
              properties:
                name:
                  type: string
                  minLength: 2
                age:
                  type: integer
                  minimum: 18
                email:
                  type: string
                  format: email
      responses:
        '201':
          description: Usuario creado
          content:
            application/json:
              schema:
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  email: 
                    type: string
                  age: 
                    type: integer
  /users/{id}:
    get:
      tags:
        - Users 
      summary: Obtener un usuario por id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name: 
                    type: string
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
    post:
      tags:
        - Users
      summary: actualizar usuario por id
      parameters:
        - in: path
          name: id
          required: true
          schema: 
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - age
                - email
              properties:
                name:
                  type: string
                  minLength: 2
                age:
                  type: integer
                  minimum: 18
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  name:
                    type: string
                  email: 
                    type: string
                  age: 
                    type: integer
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string



              